topicsフェッチ専用のカスタムフックを作成
