+++ src/types/Speaker.ts
@@ -1,0 +1,7 @@
+// todo: timestamp: any型を避けたい
+export interface Speaker {
+  id: string;
+  userId: string;
+  name: string;
+  timestamp: any;
+}
