+++ src/types/Topic.ts
@@ -1,0 +1,6 @@
+// todo: timestamp: any型を避けたい
+export interface Topic {
+  id: string;
+  content: string;
+  timestamp: any;
+}
